Add files via upload
These new code files contain the additional analyses for Denmark, Sweden, and South Australia, which have been added to the manuscript. Additionally, attached is the code with which we have implemented the analysis suggested by the editor - leaving out 2019 data and re-estimate the results. 
For the figures and tables in the revised manuscript, please go to R1_main_figures_tables_16012025.do.
For the figures and tables in the revised supplement, please go to R1_supplement_16012025.do. 
For the analysis suggested by the editor, please go to Editor_leaving2019out_31012025.do